feat & refactor: 타입 정의
- Clova 모듈 반환 타입 통일
- 서비스별 디렉토리 생성
- ClovaRequestBodyTransformer 등에서 사용할 타입 정의
