+++ src/clova/types/clova-common/clova-property.type.ts
