--- src/clova/types/clova-response.type.ts
@@ -1,3 +1,0 @@
-import { Feedback } from './feedback';
-
-export type ClovaResponse = { result: string | string[] } | Feedback[];
