+++ src/clova/types/feedback/criteria.type.ts
@@ -1,9 +1,3 @@
-export type Feedback = {
-  title: string;
-  score: string;
-  description: string;
-};
-
 export type CriteriaValue = 'TONE' | 'PURPOSE';
 
 export type CriteriaItems = { value: string; label: string };
