+++ src/clova/types/feedback/response.type.ts
@@ -1,0 +1,5 @@
+export type Feedback = {
+  title: string;
+  score: string;
+  description: string;
+};
