+++ src/clova/types/parsed-sentence/index.ts
@@ -1,2 +1,1 @@
-export * from './criteria.type';
 export * from './response.type';
