+++ src/clova/types/parsed-sentence/response.type.ts
@@ -1,0 +1,1 @@
+export type ParsedSentence = string;
