+++ src/clova/types/partial-modification/index.ts
@@ -1,2 +1,3 @@
 export * from './cache-value.type';
 export * from './command.type';
+export * from './response.type';
