+++ src/clova/types/partial-modification/response.type.ts
@@ -1,0 +1,1 @@
+export type PartialModification = { result: string | string[] };
