+++ src/clova/types/repetitive-word/index.ts
@@ -1,2 +1,1 @@
-export * from './criteria.type';
 export * from './response.type';
