+++ src/clova/types/repetitive-word/response.type.ts
@@ -1,0 +1,1 @@
+export type RepetitiveWord = string[];
