+++ src/clova/types/transform.type.ts
@@ -1,0 +1,7 @@
+import { CommandValue } from './partial-modification';
+
+export type TransformType =
+  | CommandValue
+  | 'PARSED_SENTENCE'
+  | 'REPETITIVE_WORD'
+  | 'FEEDBACK';
